+++ packages/vite/index.mjs
@@ -5,3 +5,4 @@ export * from './src/scripts.js';
 export * from './src/template-tag.js';
 export * from './src/addons.js';
 export * from './src/optimize-deps.js';
+export * from './src/build.js';
